# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lab Booking Portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExamInput {
  codiceAnalisi: string;
  descrizione: string;
  /** @nullable */
  colorProvetta?: string | null;
  synlab?: boolean;
  /** @nullable */
  um?: string | null;
  /** @nullable */
  metodo?: string | null;
  /** @nullable */
  regola?: string | null;
  /**
     * Price as decimal string
     * @nullable
     */
  importo?: string | null;
  preparationInstructions?: string;
}

export interface Exam {
  id: number;
  codiceAnalisi: string;
  descrizione: string;
  /** @nullable */
  colorProvetta?: string | null;
  synlab: boolean;
  /** @nullable */
  um?: string | null;
  /** @nullable */
  metodo?: string | null;
  /** @nullable */
  regola?: string | null;
  /**
     * Price as decimal string
     * @nullable
     */
  importo?: string | null;
  preparationInstructions?: string;
}

export interface TimeSlot {
  /** Time in HH:MM format */
  time: string;
  available: boolean;
}

/**
 * @nullable
 */
export type BookingInputGender = typeof BookingInputGender[keyof typeof BookingInputGender] | null;


export const BookingInputGender = {
  M: 'M',
  F: 'F',
} as const;

export interface BookingInput {
  /** @minItems 1 */
  examIds: number[];
  date: string;
  time: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  gender?: BookingInputGender;
  email: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
}

/**
 * @nullable
 */
export type BookingGender = typeof BookingGender[keyof typeof BookingGender] | null;


export const BookingGender = {
  M: 'M',
  F: 'F',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  accepted: 'accepted',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  examIds: number[];
  examNames: string[];
  date: string;
  time: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  gender?: BookingGender;
  email: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  createdAt: string;
  /** Number of exam results submitted for this booking */
  refertiCount?: number;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  accepted: 'accepted',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

/**
 * @nullable
 */
export type PatientGender = typeof PatientGender[keyof typeof PatientGender] | null;


export const PatientGender = {
  M: 'M',
  F: 'F',
} as const;

export interface Patient {
  id: number;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  gender?: PatientGender;
  email: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  billingAddress?: string | null;
  /** @nullable */
  billingCap?: string | null;
  /** @nullable */
  billingCity?: string | null;
  /** @nullable */
  billingProvincia?: string | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type PatientInputGender = typeof PatientInputGender[keyof typeof PatientInputGender] | null;


export const PatientInputGender = {
  M: 'M',
  F: 'F',
} as const;

export interface PatientInput {
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  gender?: PatientInputGender;
  email: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  billingAddress?: string | null;
  /** @nullable */
  billingCap?: string | null;
  /** @nullable */
  billingCity?: string | null;
  /** @nullable */
  billingProvincia?: string | null;
}

export interface Referto {
  id: number;
  bookingId: number;
  examId: number;
  valore: string;
  /** @nullable */
  note?: string | null;
  refertataAt: string;
}

export interface RefertaInput {
  bookingId: number;
  examId: number;
  valore: string;
  /** @nullable */
  note?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListSlotsParams = {
/**
 * Date in YYYY-MM-DD format
 */
date: string;
examId?: number;
};

export type ListBookingsParams = {
status?: ListBookingsStatus;
date?: string;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  pending: 'pending',
} as const;

export type ListPatientsParams = {
/**
 * Search by name, email, or phone
 */
search?: string;
};

export type ListRefertiParams = {
bookingId: number;
};