# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Lab Booking Portal
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: exams
    description: Lab exam operations
  - name: bookings
    description: Booking operations
  - name: slots
    description: Available time slot operations
  - name: patients
    description: Patient anagrafica operations
  - name: referti
    description: Exam results (refertazione)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List all available lab exams
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create a new exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /exams/{id}:
    put:
      operationId: updateExam
      tags: [exams]
      summary: Update an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "200":
          description: Exam updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /slots:
    get:
      operationId: listSlots
      tags: [slots]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format
        - name: examId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (operator use)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [confirmed, cancelled, pending]
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update the status of a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name, email, or phone
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patients/{id}:
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /referti:
    get:
      operationId: listReferti
      tags: [referti]
      summary: List exam results for a booking
      parameters:
        - name: bookingId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of exam results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Referto"
    post:
      operationId: upsertReferto
      tags: [referti]
      summary: Create or update an exam result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefertaInput"
      responses:
        "200":
          description: Referto saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ExamInput:
      type: object
      required: [codiceAnalisi, descrizione]
      properties:
        codiceAnalisi:
          type: string
        descrizione:
          type: string
        colorProvetta:
          type: ["string", "null"]
        synlab:
          type: boolean
          default: false
        um:
          type: ["string", "null"]
        metodo:
          type: ["string", "null"]
        regola:
          type: ["string", "null"]
        importo:
          type: ["string", "null"]
          description: Price as decimal string
        preparationInstructions:
          type: string
          default: ""

    Exam:
      type: object
      required: [id, codiceAnalisi, descrizione, synlab]
      properties:
        id:
          type: integer
        codiceAnalisi:
          type: string
        descrizione:
          type: string
        colorProvetta:
          type: ["string", "null"]
        synlab:
          type: boolean
        um:
          type: ["string", "null"]
        metodo:
          type: ["string", "null"]
        regola:
          type: ["string", "null"]
        importo:
          type: ["string", "null"]
          description: Price as decimal string
        preparationInstructions:
          type: string

    TimeSlot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
          description: Time in HH:MM format
        available:
          type: boolean

    BookingInput:
      type: object
      required: [examIds, date, time, firstName, lastName, dateOfBirth, email, phone]
      properties:
        examIds:
          type: array
          items:
            type: integer
          minItems: 1
        date:
          type: string
          format: date
        time:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        codiceFiscale:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: ["M", "F", null]
        email:
          type: string
          format: email
        phone:
          type: string
        notes:
          type: ["string", "null"]

    Booking:
      type: object
      required: [id, examIds, examNames, date, time, firstName, lastName, dateOfBirth, email, phone, status, createdAt]
      properties:
        id:
          type: integer
        examIds:
          type: array
          items:
            type: integer
        examNames:
          type: array
          items:
            type: string
        date:
          type: string
          format: date
        time:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        codiceFiscale:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: ["M", "F", null]
        email:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [confirmed, pending, accepted, completed, cancelled]
        createdAt:
          type: string
          format: date-time
        refertiCount:
          type: integer
          description: Number of exam results submitted for this booking

    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, pending, accepted, completed, cancelled]

    Patient:
      type: object
      required: [id, firstName, lastName, dateOfBirth, email, phone, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        codiceFiscale:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: ["M", "F", null]
        email:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        billingCap:
          type: ["string", "null"]
        billingCity:
          type: ["string", "null"]
        billingProvincia:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    PatientInput:
      type: object
      required: [firstName, lastName, dateOfBirth, email, phone]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        codiceFiscale:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: ["M", "F", null]
        email:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        billingCap:
          type: ["string", "null"]
        billingCity:
          type: ["string", "null"]
        billingProvincia:
          type: ["string", "null"]

    Referto:
      type: object
      required: [id, bookingId, examId, valore, refertataAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        examId:
          type: integer
        valore:
          type: string
        note:
          type: ["string", "null"]
        refertataAt:
          type: string
          format: date-time

    RefertaInput:
      type: object
      required: [bookingId, examId, valore]
      properties:
        bookingId:
          type: integer
        examId:
          type: integer
        valore:
          type: string
        note:
          type: ["string", "null"]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string