# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lab Booking Portal
 * OpenAPI spec version: 0.1.0
 */
import type { BookingGender } from './bookingGender';
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  examIds: number[];
  examNames: string[];
  date: Date;
  time: string;
  firstName: string;
  lastName: string;
  dateOfBirth: Date;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  gender?: BookingGender;
  email: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  createdAt: Date;
  /** Number of exam results submitted for this booking */
  refertiCount?: number;
}