# lib/api-zod/src/generated/types/referto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lab Booking Portal
 * OpenAPI spec version: 0.1.0
 */

export interface Referto {
  id: number;
  bookingId: number;
  examId: number;
  valore: string;
  /** @nullable */
  note?: string | null;
  refertataAt: Date;
}